refactor(casting): update offline_key_for function signature and use uint32_t for timing variables
- Modified the `offline_key_for` function in `net.h` to accept a size parameter for improved safety and flexibility.
- Updated timing variables in `main.cpp` and `net.h` from `unsigned long` to `uint32_t` to ensure consistency with modern C++ practices.
- These changes enhance type safety and maintainability, aligning with previous refactoring efforts in the codebase.

+++ firmware/arduino/src/main.cpp
@@ -1011,7 +1011,7 @@ void setup() {
   uint32_t display_phase_start = millis();
 // Establish a soft deadline visible to drawing helpers
 #ifdef DISPLAY_PHASE_TIMEOUT_MS
-  g_display_deadline_ms = display_phase_start + static_cast<unsigned long>(DISPLAY_PHASE_TIMEOUT_MS);
+  g_display_deadline_ms = display_phase_start + static_cast<uint32_t>(DISPLAY_PHASE_TIMEOUT_MS);
 #endif
   if (do_full) {
     full_refresh();
